% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_count.R
\name{activity_count}
\alias{activity_count}
\title{Compute activity counts using [citation] algorithm}
\usage{
activity_count(df, breaks = "5 sec", range, noise_level = 0.03, k = 0.05,
  spar = 0.6, filter_type = "butter", cutoffs = c(0.2, 5),
  integration = "trapz", axes = c(2, 3, 4), use_extrapolation = TRUE,
  use_interpolation = TRUE, use_filtering = TRUE, combination = "sum",
  vm_after_extrapolation = FALSE, before_df = NULL, after_df = NULL)
}
\arguments{
\item{df}{input sensor data frame that fits the mhealth specification}

\item{breaks}{epoch to compute counts on}

\item{range}{dynamic range of the device (two element vector)}

\item{noise_level}{noise level should be between 0.01 and 0.1}

\item{k}{neighborhood duration for extrapolation in seconds, default is 0.05 seconds}

\item{spar}{smoothing parameter for extrapolation, default is 0.6}

\item{cutoffs}{cut off frequencies to be used in filtering, default is 0.2Hz and 5Hz. If choosing bessel, the low pass cut off would be multipled by 2 when being used.}

\item{integration}{the integration method to be used: "trapz", "absoluteMeanByPoints".}

\item{resample}{set 0 to not use resampling, otherwise set to the desired sampling rate in numerical, e.g 40 for 40Hz. Default is 50.}

\item{filter}{"butter", "ellip", "bessel"}
}
\value{
Will return NA if a segment break is too short to have enough samples to compute AUC
}
