% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Calculate summary value (area under curve) for each column over a certain break (e.g. hour, min).}
\usage{
aggregate(df, breaks, type = "trapz", rectify = TRUE)
}
\arguments{
\item{df}{input dataframe that matches mhealth sensor data format.}

\item{breaks}{could be "sec", "min", "hour", "day", "week", "month", "quarter" or "year"; or preceded by an interger and a space.}

\item{type}{"trapz", "power", "sum", "meanBySecond", "meanBySize"}

\item{rectify}{whether rectify the values before computing AUC, default is TRUE}
}
\description{
Calculate summary value (area under curve) for each column over a certain break (e.g. hour, min).
}
\note{
If certain break is not provided or missing, will use the entire sequence. The column name (except for the first column) of output dataframe would be: [SUMMARY\_METHOD]\_INPUT\_HEADER\_NAME.
}
