% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{average_removal}
\alias{average_removal}
\title{Apply average removal FIR filter to the input sensor data frame each column over certain breaks (e.g. hour, sec, min and etc.)}
\usage{
average_removal(df, sr, order)
}
\arguments{
\item{df}{the input dataframe that matches mhealth specification.}

\item{sr}{sampling rate of the input signal}

\item{order}{window size (in seconds) of filter}
}
\description{
Apply average removal FIR filter to the input sensor data frame each column over certain breaks (e.g. hour, sec, min and etc.)
}
