% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{change_sampling_rate}
\alias{change_sampling_rate}
\title{Apply bandlimited interpolation filter to the input sensor data frame each column over a certain break (e.g. hour, sec, min and etc.).}
\usage{
change_sampling_rate(df, orig_sr, new_sr)
}
\arguments{
\item{orig_sr}{original sampling rate of each column}

\item{new_sr}{the desired sampling rate for each column}

\item{dft}{dataframe that matches mhealth specification.}
}
\value{
list of filtered dataframes.
}
\description{
Apply bandlimited interpolation filter to the input sensor data frame each column over a certain break (e.g. hour, sec, min and etc.).
}
