% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{compute_orientation}
\alias{compute_orientation}
\title{Calculate orientation of a sensor in (X,Y,Z degrees) for each axis}
\usage{
compute_orientation(df, epoch = 2, unit = "deg")
}
\description{
Calculate orientation of a sensor in (X,Y,Z degrees) for each axis
}
