% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cv_different_algorithms}
\alias{cv_different_algorithms}
\title{Coefficient of variation values for different acceleration data summary
algorithms}
\format{
A data frame with 30 rows and 3 variables:
\describe{
  \item{TYPE}{Accelerometer summary algorithm name, in character}
  \item{HZ}{The frequency of the elliptical shaker, in number}
  \item{COEFF_OF_VARIATION}{The coefficient of variation values, in number}
}
}
\source{
\url{https://github.com/mHealthGroup/MIMSunit-dataset-shaker/}
}
\usage{
cv_different_algorithms
}
\description{
A dataset containing the coefficient of variation values at different
frequencies for the dataset that includes accelerometer measures of different
devices on a standard elliptical shaker.
}
\keyword{datasets}
