% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate.R
\name{extrapolate.data.frame}
\alias{extrapolate.data.frame}
\title{Apply extrapolate algorithm to a single stream of data.}
\usage{
extrapolate.data.frame(df, ...)
}
\arguments{
\item{t}{input index or timestamp sequence}

\item{k}{duration of neighborhood to be used in local spline regression for each side, in seconds}

\item{spar}{between 0 and 1, to control how smooth we want to fit local spline regression, 0 is linear and 1 matches all local points. A good choice is 0.3 to penalize the maxed out points.}
}
\note{
If wanting to obtain mediate results for plotting and debugging, please use this function.
}
\author{
Qu Tang
}
