% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{iir}
\alias{iir}
\title{Apply iir filter to the input sensor data frame each column over a certain break (e.g. hour, sec, min and etc.).}
\usage{
iir(df, Fs, Fc, order, type = "high", filter_type = "butter")
}
\arguments{
\item{df}{the input dataframe that matches mhealth specification.}

\item{Fs}{sampling rate of the input signal}

\item{Fc}{cut off frequencies of butterworth filter, if more than one store as c(low, high)}

\item{order}{formula order of butterworth filter}

\item{type}{"low", "high", "stop", "pass"}

\item{filter_type}{"butter", "chebyI", "chebyII", "ellip"}
}
\value{
list of filtered dataframes.
}
\description{
Apply iir filter to the input sensor data frame each column over a certain break (e.g. hour, sec, min and etc.).
}
