% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{illustrate_extrapolation}
\alias{illustrate_extrapolation}
\title{Plot illustrations about extrapolation in illustration style.}
\usage{
illustrate_extrapolation(df, between_neighbor_df = NULL, title = NULL,
  left_neighbors_df = NULL, right_neighbors_df = NULL,
  show_neighbors = T, extrap_points_df = NULL, fitted_line_df = NULL,
  show_extrapolated_points_and_lines = T)
}
\arguments{
\item{df}{data.frame.The original data before extrapolation.}

\item{between_neighbor_df}{data.frame. Dataframe containing points between the points used for extrapolation.}

\item{left_neighbors_df}{data.frame. Dataframe containing the points of left side used for extrapolation.}

\item{right_neighbors_df}{data.frame. Dataframe containing the points of right side used for extrapolation.}

\item{show_neighbors}{bool. Show the points used for extrapolation if TRUE.}

\item{extrap_points_df}{data.frame. Dataframe containing the points got from extrapolation.}

\item{fitted_line_df}{data.frame. Dataframe containing the spline curves used during extrapolation.}

\item{show_extrapolated_points_and_lines}{bool. Show the extrapolated points and curves used for extrapolation.}
}
\value{
ggplot2 graph object. The graph to be shown.
}
\description{
\code{illustrate_extrapolation} plots elements of extrapolations
(e.g., marked points, reference lines) in the same style as
\code{\link{illustrate_signal}}.
}
\seealso{
Other visualization functions.: \code{\link{illustrate_signal}}
}
\concept{visualization functions.}
