% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_actigraph_count}
\alias{import_actigraph_count}
\title{Import and convert Actigraph count csv files and load into data frame as in mhealth format.}
\usage{
import_actigraph_count(filename, col_name = "ACTIGRAPH_COUNT", axes = c(2,
  3, 4))
}
\arguments{
\item{filename}{full file path of input Actigraph count csv file.}
}
