% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_actigraph_count_vm}
\alias{import_actigraph_count_vm}
\title{Import and convert Actigraph count csv files and load into data frame as in mhealth format.}
\usage{
import_actigraph_count_vm(filename, col_name = "ACTIGRAPH_COUNT")
}
\arguments{
\item{filename}{full file path of input Actigraph count csv file.}
}
