% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_actigraph_csv_chunked}
\alias{import_actigraph_csv_chunked}
\title{Import large raw multi-channel accelerometer data stored in Actigraph raw csv
format in chunks}
\usage{
import_actigraph_csv_chunked(
  filepath,
  in_voltage = FALSE,
  has_ts = TRUE,
  header = TRUE
)
}
\arguments{
\item{filepath}{string. The filepath of the input data.}

\item{in_voltage}{set as TRUE only when the input Actigraph csv file is in
analog quantized format and need to be converted into g value}

\item{has_ts}{set as TRUE only when timestamp is provided as the first column}

\item{header}{boolean. If TRUE, the input csv file will have column names in
the first row.}
}
\value{
list. The list contains three items. The first item is a generator
function that each time it is called, it will
return a list of `list(before_df, df, after_df)`. These are data.frames of
imported chunks, ordered in time that can be loaded with
the `MIMSunit::mims_unit` function directly. The second item is a
`has_more_chunks` function used to check if all chunks are loaded.
If it returns `FALSE`, it means the loading has ended. The third item is a
`close` function which you can call at any moment to close the file loading.
}
\description{
\code{import_actigraph_csv_chunked} imports the raw multi-channel accelerometer data
stored in Actigraph raw csv format. It supports files from the following
devices: GT3X, GT3X+, GT3X+BT, GT9X, and GT9X-IMU.
}
\details{
For old device (GT3X) that stores accelerometer values as digital voltage.
The function will convert the values to \eqn{g} unit using the following
equation.

\deqn{x_g = \frac{x_{voltage}r}{(2 ^ r) - \frac{v}{2}}}

Where \eqn{v} is the max voltage corresponding to the max accelerometer value
that can be found in the meta section in the csv file; \eqn{r} is the
resolution level which is the number of bits used to store the voltage
values. \eqn{r} can also be found in the meta section in the csv file.
}
\section{How is it used in MIMS-unit algorithm?}{
 This function is a File IO
  function that is used to import data from Actigraph devices during
  algorithm validation.
}

\seealso{
Other Filo I/O functions: 
\code{\link{export_to_actilife}()},
\code{\link{import_actigraph_count_csv}()},
\code{\link{import_actigraph_csv}()},
\code{\link{import_actigraph_meta}()},
\code{\link{import_activpal3_csv}()},
\code{\link{import_enmo_csv}()},
\code{\link{import_mhealth_csv_chunked}()},
\code{\link{import_mhealth_csv}()}
}
\concept{Filo I/O functions}
