% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_actigraph_meta}
\alias{import_actigraph_meta}
\title{parse actigraph csv header to get related version and sampling rate information}
\usage{
import_actigraph_meta(filename, header = TRUE)
}
\description{
parse actigraph csv header to get related version and sampling rate information
}
