% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_activpal}
\alias{import_activpal}
\alias{import_activpal_raw}
\title{Import ActivPal Raw data files and load into dataframe as mhealth format.}
\usage{
import_activpal_raw(filename, header_provided = FALSE)
}
\description{
Import ActivPal Raw data files and load into dataframe as mhealth format.
}
