% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_biobank_enmo}
\alias{import_biobank_enmo}
\title{Import and convert biobank epoch csv files and load into data frame as in mhealth format.}
\usage{
import_biobank_enmo(filepath, col_name = "biobank_enmo")
}
\arguments{
\item{filepath}{full file path of input biobank epoch csv file.}
}
\description{
Import and convert biobank epoch csv files and load into data frame as in mhealth format.
}
