% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_mhealth_csv}
\alias{import_mhealth_csv}
\title{Import raw multi-channel accelerometer data stored in mHealth Specification}
\usage{
import_mhealth_csv(filepath)
}
\arguments{
\item{filepath}{string. The filepath of the input data.}
}
\value{
dataframe. The imported multi-channel accelerometer signal, with the
  first column being the timestamps in POSXlct format, and the rest columns
  being accelerometer values in \eqn{g} unit.
}
\description{
\code{import_mhealth_csv} imports the raw multi-channel accelerometer data
stored in mHealth Specification.
}
\section{How is it used in MIMS-unit algorithm?}{
 This function is a File IO
  function that is used to import data stored in mHealth Specification during
  algorithm validation.
}

\seealso{
Other File I/O functions: 
\code{\link{export_to_actilife}()},
\code{\link{import_actigraph_count_csv}()},
\code{\link{import_actigraph_csv_chunked}()},
\code{\link{import_actigraph_csv}()},
\code{\link{import_actigraph_meta}()},
\code{\link{import_activpal3_csv}()},
\code{\link{import_enmo_csv}()},
\code{\link{import_mhealth_csv_chunked}()}
}
\concept{File I/O functions}
