% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_mhealth_csv_chunked}
\alias{import_mhealth_csv_chunked}
\title{Import large raw multi-channel accelerometer data stored in mHealth Specification
in chunks.}
\usage{
import_mhealth_csv_chunked(filepath)
}
\arguments{
\item{filepath}{string. The filepath of the input data.}
}
\value{
list. The list contains three items. The first item is a generator
function that each time it is called, it will
return a list of `list(before_df, df, after_df)`. These are data.frames of
imported chunks, ordered in time that can be loaded with
the `MIMSunit::mims_unit` function directly. The second item is a
`has_more_chunks` function used to check if all chunks are loaded.
If it returns `FALSE`, it means the loading has ended. The third item is a
`close` function which you can call at any moment to close the file loading.
}
\description{
\code{import_mhealth_csv_chunked} imports the raw multi-channel accelerometer
data stored in mHealth Specification in chunks.
}
\section{How is it used in MIMS-unit algorithm?}{
 This function is a File IO
  function that is used to import data stored in mHealth Specification during
  algorithm validation.
}

\seealso{
Other Filo I/O functions: 
\code{\link{export_to_actilife}()},
\code{\link{import_actigraph_count_csv}()},
\code{\link{import_actigraph_csv_chunked}()},
\code{\link{import_actigraph_csv}()},
\code{\link{import_actigraph_meta}()},
\code{\link{import_activpal3_csv}()},
\code{\link{import_enmo_csv}()},
\code{\link{import_mhealth_csv}()}
}
\concept{Filo I/O functions}
