% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate the missing points and unify sampling interval for the input sensor data}
\usage{
interpolate(df, method = "spline_natural", sr = 100, st = NULL,
  et = NULL, polyDegree = 3)
}

