% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_signal}
\alias{interpolate_signal}
\title{Interpolate the missing points and unify sampling interval for the input sensor data}
\usage{
interpolate_signal(df, method = "spline_natural", sr = 100,
  st = NULL, et = NULL)
}
\description{
Interpolate the missing points and unify sampling interval for the input sensor data
}
