% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{magnitude}
\alias{magnitude}
\title{Compute the magnitude value of sensor data.}
\usage{
magnitude(df, axes = NULL)
}
\arguments{
\item{df}{input dataframe that matches specification.}
}
\description{
Compute the magnitude value of sensor data.
}
