% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measurements_different_devices}
\alias{measurements_different_devices}
\title{The mean and standard deviation of accelerometer summary measures for
different acceleration data summary algorithms and for different devices.}
\format{A data frame with 235 rows and 8 variables:
\describe{
  \item{DEVICE}{The name of different devices, in character}
  \item{GRANGE}{The dynamic range of the device in g, in number}
  \item{SR}{The sampling rate in Hz of the device, in number}
  \item{TYPE}{Accelerometer summary algorithm name, in character}
  \item{HZ}{The frequency of the elliptical shaker, in number}
  \item{NAME}{An alternative name that is friendly for plotting for devices, in character}
  \item{mean}{The mean values of accelerometer summary measure, in number}
  \item{sd}{The standard deviation values of accelerometer summary measure, in number}
}}
\source{
\url{https://github.com/qutang/MIMSunit-dataset-shaker}
}
\usage{
measurements_different_devices
}
\description{
A dataset containing the mean and standard deviation of acceleometer summary
measures at different frequencies for the dataset that includes accelerometer
measures of different devices on a standard ellpitical shaker.
}
\keyword{datasets}
