% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{parse_epoch_string}
\alias{parse_epoch_string}
\title{Parse epoch string to the corresponding number of samples it represents.}
\usage{
parse_epoch_string(epoch_str, sr)
}
\arguments{
\item{epoch_str.}{string. The input epoch str as accepted by \code{breaks}
argument of \code{\link[base]{cut.POSIXt}}.}

\item{sr.}{number. The sampling rate in Hz used to parse the epoch string.}
}
\value{
number. The number of samples represented by the epoch string.
}
\description{
\code{parse_epoch_string} parses the epoch string (e.g. "1 min"), and outputs
the corresponding number of samples it represents.
}
\details{
This function parses the given epoch string (e.g. "5 secs") and outputs the
corresponding number of samples represented by the epoch string.
}
\section{How is it used in MIMS-unit algorithm?}{
 This function is used in
  \code{\link{aggregate_for_mims}} function.
}

\seealso{
Other utility functions: \code{\link{cut_off_signal}},
  \code{\link{interpolate_signal}},
  \code{\link{sampling_rate}},
  \code{\link{simulate_new_data}}
}
\concept{utility functions}
