% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{resample}
\alias{resample}
\title{Apply bandlimited interpolation filter to the input sensor data frame each column over a certain break (e.g. hour, sec, min and etc.).}
\usage{
resample(df, origSr, newSr)
}
\arguments{
\item{origSr}{original sampling rate of each column}

\item{newSr}{the desired sampling rate for each column}

\item{dft}{dataframe that matches mhealth specification.}
}
\value{
list of filtered dataframes.
}

