% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rest_on_table}
\alias{rest_on_table}
\title{A short snippet of raw accelerometer signal from a device resting on a table.}
\format{
A data frame with 5000 rows and 4 variables:
\describe{
  \item{HEADER_TIME_STAMP}{The timestamp of raw accelerometer data, in POSIXct}
  \item{X}{The x axis value of raw accelerometer data, in number}
  \item{Y}{The x axis value of raw accelerometer data, in number}
  \item{Z}{The x axis value of raw accelerometer data, in number}
}
}
\source{
\url{https://github.com/mHealthGroup/MIMSunit/}
}
\usage{
rest_on_table
}
\description{
The dataset includes accelerometer data sampled at 80Hz and 6g. This data is used to derive the thresholding.
}
\keyword{datasets}
