% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_raw_accel_data}
\alias{sample_raw_accel_data}
\title{Sample raw accelerometer data}
\format{
A data frame with 480 rows and 4 variables:
\describe{
  \item{HEADER_TIME_STAMP}{Timestamp, in POSIXct}
  \item{X}{X axis values, in number}
  \item{Y}{Y axis values, in number}
  \item{Z}{Z axis values, in number}
}
}
\source{
\url{https://github.com/mHealthGroup/MIMSunit/}
}
\usage{
sample_raw_accel_data
}
\description{
A raw accelerometer data file contains treadmill data collected from a human
subject.
}
\keyword{datasets}
