% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{segment_data}
\alias{segment_data}
\title{Segment input dataframe into windows as specified by breaks.
\code{segment_data} segments the input sensor dataframe into
 epoch windows with length specified in breaks.}
\usage{
segment_data(df, breaks)
}
\arguments{
\item{df}{dataframe. Input dataframe of the multi-channel signal. The first
column is the timestamps in POSXlct format and the following columns are
accelerometer values.}

\item{breaks}{character. An epoch length character that can be accepted by
cut.breaks function.}
}
\value{
dataframe. The same format as the input dataframe, but with an extra
column "SEGMENT" in the end specifies the epoch window a sample belongs to.
}
\description{
This function accepts a dataframe of multi-channel signal, segments it
into epoch windows with length specified in breaks.
}
\section{How is it used in MIMS-unit algorithm?}{
 This function is a utility
  function that was used in various part in the algorithm whenever we need to
  segment a dataframe, e.g., before aggregating values over epoch windows.
}

\seealso{
Other utility functions: 
\code{\link{clip_data}()},
\code{\link{cut_off_signal}()},
\code{\link{interpolate_signal}()},
\code{\link{parse_epoch_string}()},
\code{\link{sampling_rate}()},
\code{\link{simulate_new_data}()}
}
\concept{utility functions}
