% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{shiny_app}
\alias{shiny_app}
\title{Run shiny app to compute MIMSunit values from files}
\usage{
shiny_app(options = list())
}
\arguments{
\item{options}{The options passed to \code{\link[shiny]{shinyApp}}.}
}
\description{
\code{shiny_app} runs a local shiny app that provides a user friendly interface
to compute mims unit values from files stored in mhealth or actigraph format.
}
\section{How is it used in MIMS-unit algorithm?}{
 This provides a user friendly
graphical interface to load local files, call \code{\link{mims_unit_from_files}}
and display the results as an interactive graph.
}

\examples{
  \dontrun{
  shiny_app()
  }
}
\seealso{
Other Top level API functions: 
\code{\link{custom_mims_unit}()},
\code{\link{mims_unit}()},
\code{\link{sensor_orientations}()}
}
\concept{Top level API functions}
