% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_axes.R
\name{sum_up}
\alias{sum_up}
\title{Compute the sum up value of sensor data.}
\usage{
sum_up(df, axes = NULL)
}
\arguments{
\item{df}{input dataframe that matches specification.}
}
\description{
Compute the sum up value of sensor data.
}
