% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{sumUp}
\alias{sumUp}
\title{Compute the sumUp value of sensor data.}
\usage{
sumUp(df, axes = NULL)
}
\arguments{
\item{df}{input dataframe that matches specification.}
}
\description{
Compute the sumUp value of sensor data.
}
