% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_axes.R
\name{vector_magnitude}
\alias{vector_magnitude}
\title{Compute the vector magnitude value of sensor data.}
\usage{
vector_magnitude(df, axes = NULL)
}
\arguments{
\item{df}{input dataframe that matches specification.}
}
\description{
Compute the vector magnitude value of sensor data.
}
