% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_over_epoch.R
\name{aggregate_for_orientation}
\alias{aggregate_for_orientation}
\title{Calculate orientation value for each column over a certain break (e.g. hour, min).}
\usage{
aggregate_for_orientation(df, breaks, epoch = 2, unit = "deg")
}
\arguments{
\item{df}{input dataframe that matches mhealth sensor data format.}

\item{breaks}{could be 'sec', 'min', 'hour', 'day', 'week', 'month', 'quarter' or 'year'; or preceded by an interger and a space.}
}
\description{
Calculate orientation value for each column over a certain break (e.g. hour, min).
}
\note{
If certain break is not provided or missing, will use the entire sequence. The column name (except for the first column) of output dataframe would be: [SUMMARY\_METHOD]\_INPUT\_HEADER\_NAME.
}
