% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measurements_different_devices}
\alias{measurements_different_devices}
\title{The mean and standard deviation of accelerometer summary measure for
different acceleration data summary algorithms and for different devices.}
\format{
A data frame with 235 rows and 8 variables:
\describe{
  \item{DEVICE}{The name of different devices, in character}
  \item{GRANGE}{The dynamic range of the device in g, in number}
  \item{SR}{The sampling rate in Hz of the device, in number}
  \item{TYPE}{Accelerometer summary algorithm name, in character}
  \item{HZ}{The frequency of the elliptical shaker, in number}
  \item{NAME}{An alternative name that is friendly for plotting for devices, in character}
  \item{mean}{The mean values of accelerometer summary measure, in number}
  \item{sd}{The standard deviation values of accelerometer summary measure, in number}
}
}
\source{
\url{https://github.com/mHealthGroup/MIMSunit-dataset-shaker/}
}
\usage{
measurements_different_devices
}
\description{
A dataframe contains the mean and standard deviation of accelerometer summary
measured at different frequencies for the raw accelerometer signals
from different devices collected from on a standard elliptical shaker.
}
\keyword{datasets}
